In PluginHost.js testing for existence of Plugin.name before publishingh anf firing:
IE chocked because apparently sometimes´the Plugin.name is null. This allows IE to load the page and
Clicking on dialogues works the 2 time, then it chokes again due to missing NS.
The first click on the app button does nothing. The second opens the dialogue and the third fails.